robo-advisor-ai/backend/requirements.txt
# Step 1: Data Models + Optimizer
numpy>=1.24.0
scipy>=1.10.0
pydantic>=2.0.0

# Step 2: Supabase Memory Layer
supabase>=2.0.0
python-dotenv>=1.0.0

# Step 3: Market Data Tools
yfinance>=0.2.36
requests>=2.31.0
anthropic>=0.39.0

# Step 4: Agent Graph
langgraph>=0.2.0

# Step 5: FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
